+++ internal/pkg/dbmigrate/migrations/000050_alert_event_datasource_id.down.sql
@@ -1,0 +1,2 @@
+ALTER TABLE alert_events DROP INDEX idx_alert_events_datasource_id;
+ALTER TABLE alert_events DROP COLUMN datasource_id;